refactoring SCBackendService to support Exchange as a backend
