+++ spatialconnect/src/main/java/com/boundlessgeo/spatialconnect/SpatialConnect.java
@@ -22,6 +22,8 @@ import com.boundlessgeo.spatialconnect.scutilities.SCCache;
 import com.boundlessgeo.spatialconnect.services.SCBackendService;
 import com.boundlessgeo.spatialconnect.services.SCConfigService;
 import com.boundlessgeo.spatialconnect.services.SCDataService;
+import com.boundlessgeo.spatialconnect.services.SCExchangeBackendService;
+import com.boundlessgeo.spatialconnect.services.SCMqttBackendService;
 import com.boundlessgeo.spatialconnect.services.SCSensorService;
 import com.boundlessgeo.spatialconnect.services.SCService;
 import com.boundlessgeo.spatialconnect.services.SCServiceGraph;
@@ -59,7 +61,8 @@ public class SpatialConnect {
 
     private SpatialConnect() {}
 
-    /** Singleton of SpatialConnect
+    /**
+     * Singleton of SpatialConnect
      *
      * @return instance of the SpatialConnect object singleton
      */
@@ -85,6 +88,7 @@ public class SpatialConnect {
     /**
      * Adds an instantiated instance of Service that extends the {@link SCService} class.
      * The 'service' must extend the {@link SCService} class.
+     *
      * @param service the service object
      */
     public void addService(SCService service) {
@@ -93,6 +97,7 @@ public class SpatialConnect {
 
     /**
      * This stops and removes a {@link SCService} from the SpatialConnect instance.
+     *
      * @param serviceId the id of the service to be removed
      */
     public void removeService(String serviceId) {
@@ -104,6 +109,7 @@ public class SpatialConnect {
      *
      * Do not call service start on the service instance. Use this method to
      * start a {@link SCService}.
+     *
      * @param serviceId the id of the service that needs to start
      */
     public void startService(final String serviceId) {
@@ -125,6 +131,7 @@ public class SpatialConnect {
      *
      * Do not call service stop on the service instance. Use this method to
      * stop a {@link SCService}.
+     *
      * @param serviceId the id of the service that needs to be stopped.
      */
     public void stopService(String serviceId) {
@@ -145,6 +152,7 @@ public class SpatialConnect {
      *
      * Do not call service restart on the service instance. Use this method
      * to start a {@link SCService}.
+     *
      * @param serviceId the id of the service that needs to be restarted.
      */
     public void restartService(String serviceId) {
@@ -161,6 +169,7 @@ public class SpatialConnect {
 
     /**
      * Finds service from service id
+     *
      * @param serviceId the id of the service that needs to be retrieved
      * @return instance of service found by id
      */
@@ -204,15 +213,19 @@ public class SpatialConnect {
     /**
      * If you have an instance of SpatialConnect Server, this is how you would register it.
      * Passing in a remote configuration object will use the info to start the connection to the backend.
-     * @param remoteConfig object that represents http and mqtt endpoints {@link SCRemoteConfig}.
+     *
+     * @param remoteConfig object that contains http and/or mqtt endpoints {@link SCRemoteConfig}.
      */
     public void connectBackend(final SCRemoteConfig remoteConfig) {
-        if (serviceGraph.getNodeById(SCBackendService.serviceId()) == null) {
-            backendService = new SCBackendService(context);
-            backendService.initialize(remoteConfig);
-            addService(backendService);
-            startService(backendService.getServiceId());
+        Log.v(LOG_TAG, "connecting to backend");
+        if (remoteConfig.getAuth().equalsIgnoreCase("exchange")) {
+            backendService = new SCExchangeBackendService(context);
+        } else {
+            backendService = new SCMqttBackendService(context);
         }
+        backendService.initialize(remoteConfig);
+        addService(backendService);
+        startService(backendService.getServiceId());
     }
 
     public void connectAuth(ISCAuth authMethod) {
@@ -225,6 +238,11 @@ public class SpatialConnect {
         }
     }
 
+    /**
+     * This method is used by an app to send its Firebase token.
+     *
+     * @param token
+     */
     public void updateDeviceToken(final String token) {
 
         serviceRunning(SCBackendService.serviceId()).subscribe(
@@ -237,7 +255,7 @@ public class SpatialConnect {
                 }, new Action0() {
                     @Override
                     public void call() {
-                        backendService.updateDeviceToken(token);
+                        ((SCMqttBackendService) backendService).updateDeviceToken(token);
                     }
                 });
     }
@@ -269,6 +287,7 @@ public class SpatialConnect {
     /**
      * Identifies an application instance running on a device. This is the recommended
      * solution for non-ads use-cases.
+     *
      * @return String UUID string of the install id
      */
     public String getDeviceIdentifier() {
